[simpler-hooks] merged the types into the hook moved hook into Hooks
